// src/app/_modules/product/models/product.model.ts
export class Product {
    id?: any;
    itemname?: String;
    description?:  String;
    images?:  any[];
    design?:  String;
    manufacturer?: String;
    cpu?:  String;
    memory?:  String;
    rom?: String;
    config?:  String;
    display?: String;
    stock?:  String;
    unitprice?: String;
    discount?: String;
    shippingprice?:  String;
    shippingdetail?:  String;
    deliverytime?: String;
    published?: boolean;
}